+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
 ### Added
 
 - New filter: `pretty-xml` to indent/pretty-print XML documents
-- New reporter: '[prowl](https://prowlapp.com)' (by nitz).
+- New reporter: `prowl` (by nitz)
 
 ### Fixed